Extract methods to interfaces & change way of providing dependencies (#25)
* Extract methods to interfaces and provide dependencies

* Provide preferences and replace singleton

* Create module

* Extract methods to interfaces

* Replace singleton

* Use callback instead of using singleton

* Refactor and hook callback

* Rename method
